+++ samples/Pdf/Mpdf2.php
@@ -34,9 +34,9 @@ class Mpdf2 extends Mpdf
         $fontFile = 'ShadowsIntoLight-Regular.ttf';
         $config['fontdata'] = $fontdata + [ // lowercase letters only in font key
             'shadowsintolight' => [
-                    'R' => $fontFile,
-                ],
-            ];
+                'R' => $fontFile,
+            ],
+        ];
 
         return new \Mpdf\Mpdf($config);
     }